LayoutTests/imported/w3c/web-platform-tests/css/css-animations/revert-rule-keyframes-dynamic-expected.txt


FAIL revert-rule in keyframes, reverted-to value changing (standard property) assert_equals: expected "160px" but got "150px"
FAIL revert-rule in keyframes, reverted-to value changing (custom property) assert_equals: expected "160px" but got "150px"
